+++ platform-tests/src/test/java/org/junit/platform/launcher/listener/SummaryGenerationTests.java
@@ -18,6 +18,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
@@ -29,6 +30,7 @@ import org.junit.platform.engine.test.TestDescriptorStub;
 import org.junit.platform.launcher.TestIdentifier;
 import org.junit.platform.launcher.TestPlan;
 import org.junit.platform.launcher.listeners.SummaryGeneratingListener;
+import org.junit.platform.launcher.listeners.TestExecutionSummary;
 
 /**
  * @since 1.0
@@ -142,9 +144,10 @@ class SummaryGenerationTests {
 		listener.executionStarted(failingTest);
 		listener.executionFinished(failingTest, TestExecutionResult.failed(failedException));
 		listener.testPlanExecutionFinished(testPlan);
-		assertThat(listener.getSummary().getFailures()).hasSize(1);
-		assertThat(listener.getSummary().getFailures().get(0).getException()).isEqualTo(failedException);
-		assertThat(listener.getSummary().getFailures().get(0).getTestIdentifier()).isEqualTo(failingTest);
+		final List<TestExecutionSummary.Failure> failures = listener.getSummary().getFailures();
+		assertThat(failures).hasSize(1);
+		assertThat(failures.get(0).getException()).isEqualTo(failedException);
+		assertThat(failures.get(0).getTestIdentifier()).isEqualTo(failingTest);
 	}
 
 	@Test
